requirements.txt
fastapi
uvicorn[standard] 
sqlalchemy
psycopg[binary]
gymnasium
numpy
psycopg2-binary
simpy
python-dotenv
